requirements.txt
flask
PyPDF2
python-docx
textract
ollama
